+++ src/test/resources/org/jenkinsci/plugins/kubernetes/credentials/mockedKubectl.groovy
@@ -1,0 +1,8 @@
+node{
+  label "mocked-kubectl"
+  stage('Run') {
+    withKubeConfig([credentialsId: 'cred1234', serverUrl: 'https://localhost:6443']) {
+      echo "File has been configured ${env.KUBECONFIG}"
+    }
+  }
+}