+++ CHANGELOG.md
@@ -5,7 +5,8 @@ CHANGELOG
 -----
 * Import authentication feature from the kubernetes plugin
 * Add support for StringCredentials
+* Add support for scoped credentials
+* Add support for tokens and password with spaces
 * Embed certificates into the kubeconfig
 * Fix the masks used on the commands to prevent tokens from leaking into the logs
-* Add support for scoped credentials
 * Display kubectl output in the logs